requirements.txt
# FIBA AI — All Dependencies
# ============================
# Run: pip install -r requirements.txt

# Core web framework (Yash)
flask>=2.3.0
flask-cors>=4.0.0

# Computer vision
opencv-python-headless>=4.8.0

# Hand detection (Atul)
mediapipe>=0.10.0

# Object detection (Atul)
ultralytics>=8.0.0

# Numerics + signal processing (Tanishk: motion engine)
numpy>=1.24.0
scipy>=1.11.0

# Image utilities
Pillow>=9.0.0

# NLP for query parser (Atul — optional, falls back to regex)
# spacy>=3.8.0
# Run after: python -m spacy download en_core_web_sm

# Optional: MobileSAM for enhanced segmentation (Tanishk)
# mobile-sam

# Recommended: CLIP for semantic action verification (huge accuracy boost)
# Requires PyTorch first, then one of the CLIP packages.
# The pipeline works without these but accuracy is significantly better with CLIP.
torch>=2.0.0
torchvision>=0.15.0

# Option A: OpenAI CLIP (simpler)
# pip install git+https://github.com/openai/CLIP.git

# Option B: open-clip-torch (easier to install)
open-clip-torch>=2.20.0

